Move parameter functions to util
