+++ v-next/core/src/types/common.ts
@@ -44,38 +44,3 @@ export type ParameterValue =
  */
 export type ParameterTypeToValueType<T extends ParameterType> =
   ParameterToValueTypeMap[T];
-
-const parameterTypeValidators: Record<
-  ParameterType,
-  (value: unknown) => boolean
-> = {
-  [ParameterType.STRING]: (value): value is string => typeof value === "string",
-  [ParameterType.BOOLEAN]: (value): value is boolean =>
-    typeof value === "boolean",
-  [ParameterType.INT]: (value): value is number => Number.isInteger(value),
-  [ParameterType.BIGINT]: (value): value is bigint => typeof value === "bigint",
-  [ParameterType.FLOAT]: (value): value is number => typeof value === "number",
-  [ParameterType.FILE]: (value): value is string => typeof value === "string",
-};
-
-/**
- * Checks if a parameter value is valid for a given parameter type.
- *
- * This function uses a map of validators, where each validator is a function
- * that checks if a value is valid for a specific parameter type.
- * If the parameter type is variadic, the value is considered valid if it is an
- * array and all its elements are valid for the parameter type.
- */
-export function isParameterValueValid(
-  type: ParameterType,
-  value: unknown,
-  isVariadic: boolean = false,
-): boolean {
-  const validator = parameterTypeValidators[type];
-
-  if (isVariadic) {
-    return Array.isArray(value) && value.every(validator);
-  }
-
-  return validator(value);
-}
